# Improve performance and remove str limit