--- a/index.js
+++ b/index.js
@@ -50,9 +50,7 @@ function parse(str) {
   if (str.length > 100) {
     return;
   }
-  var match = /^((?:\d+)?\-?\d?\.?\d+) *(milliseconds?|msecs?|ms|seconds?|secs?|s|minutes?|mins?|m|hours?|hrs?|h|days?|d|weeks?|w|years?|yrs?|y)?$/i.exec(
-    str
-  );
+  var match = /^(-?(?:\d+)?\.?\d+) *(\w*)$/i.exec(str);
   if (!match) {
     return;
   }