@Delagen 2017-06-07
Strange how you get this results.
My:
original x 94,138 ops/sec ±0.43% (91 runs sampled)
new x 95,883 ops/sec ±0.86% (90 runs sampled)
thevtm x 92,011 ops/sec ±1.15% (88 runs sampled)
Fastest is new