@thevtm 2017-06-07
@Delagen 
Well the hottest piece of code that changed was the Regex and since you updated it in your version both implementations are basically equal in terms of performance. 
Also the original implementation cheats on  `ms('3'.repeat(998) + 'd');
  ms('4'.repeat(998) + 'x');`.
The rest is different Node version (using v8.0.0) and hardware.
```
original x 39,908 ops/sec ±1.04% (86 runs sampled)
new x 73,619 ops/sec ±2.29% (85 runs sampled)
mine x 73,813 ops/sec ±1.79% (87 runs sampled)
Fastest is mine
Done in 18.37s.
```