@sam-github 2017-06-08
I'm not sure which of these improvements will eventually merge, but when/if it does, could you release it as a patch update to the 0.7.x release line, as well as on the 2.x? That would help a lot of projects that are stuck on 0.7, possibly via deep dependency trees.