@Delagen 2017-06-13
@sam-github I think author must create new branch from 0.7 tag to PR to it