@TooTallNate 2017-11-29
Why remove the string length limit?