@thevtm 2017-11-29
Mostly because it wasn't there before the performance issues.