@TooTallNate 2017-11-29
Well it was added because of a ReDoS attack, so we probably shouldn't remove it: https://github.com/zeit/ms/commit/caae2988ba2a37765d055c4eee63d383320ee662