@sam-github 2017-11-29
The redos was possible because of an expensive regex over unconstrained length input, notice the replacement regex doesn't allow unconstrained length input.