@thevtm 2017-11-30
I've been running some benchmarks and I've found an issue with the current proposed regex.
I'm also in favor of keeping the string limit, JS regex engine is really weird.
I'll update the PR.