@Delagen 2017-11-30
Lol guys. There is zeit/ms#93 where was long conversation about limit and regex, it still not merged and without any activity for about half year