@thevtm 2017-11-30
I've finished the benchmarking and came up with a faster and more robust regex rule for parse.
[Benchmark results and code](https://gist.github.com/thevtm/357befbd285eda3bab635f20f43a2563)
Updated branch:
* New regex
* Kept str limit